move outgoing traffic normalization helper function used in tests to vcr.utils
